+++ apps/api/src/plugins/internal/authentication.ts
@@ -1,11 +1,9 @@
-import FastifyBetterAuth, {
-  type FastifyBetterAuthOptions,
-} from 'fastify-better-auth';
+import FastifyBetterAuth, { type FastifyBetterAuthOptions } from 'fastify-better-auth';
 
 import auth from '#/auth.js';
 
 export const autoConfig: FastifyBetterAuthOptions = {
-  auth,
+  auth
 };
 
 export default FastifyBetterAuth;