+++ packages/logger/src/index.ts
@@ -9,9 +9,9 @@ export const loggerOptions: LoggerOptions = {
       : {
           target: 'pino-pretty',
           options: {
-            colorize: true,
-          },
-        },
+            colorize: true
+          }
+        }
 };
 
 export const logger: Logger = pino(loggerOptions);